// types/types.ts
export interface ChannelDetails {
    title: string;
    description: string;
    thumbnail: string;
    channelId: string;
}

export interface VideoDetails {
    title: string;
    description: string;
    thumbnail: string;
    videoId: string;
    channelTitle: string;
    publishedAt: string;
    duration: string;
    viewCount: string;
    likeCount: string;
    views: string;
    likes: string;
    comments: any;
    tags: string[];
    channel: any;
 
}